refactor steps into separate files
