+++ recipes/food-ordering/android/food-ordering/src/main/java/ai/picovoice/foodordering/Steps/Step.java
@@ -1,0 +1,26 @@
+/*
+    Copyright 2026 Picovoice Inc.
+
+    You may not use this file except in compliance with the license. A copy of the license is
+    located in the "LICENSE" file accompanying this source.
+
+    Unless required by applicable law or agreed to in writing, software distributed under the
+    License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+    express or implied. See the License for the specific language governing permissions and
+    limitations under the License.
+*/
+
+package ai.picovoice.foodordering.Steps;
+
+import ai.picovoice.foodordering.AINoiseSuppressedRecorder;
+import ai.picovoice.foodordering.WorkflowListener;
+
+public abstract class Step {
+    protected AINoiseSuppressedRecorder recorder;
+    protected WorkflowListener listener;
+    public Step(AINoiseSuppressedRecorder recorder, WorkflowListener listener) {
+        this.recorder = recorder;
+        this.listener = listener;
+    }
+    public abstract void delete();
+}
